requirements.txt
fastapi>=0.115.0
uvicorn>=0.27.0
python-multipart>=0.0.9
langchain>=0.1.0
langchain-google-genai>=0.0.5
PyPDF2>=3.0.0
python-dotenv>=1.0.0